Move ChatItemData model and ChatItem widget to separate file
